aufs: support for /proc/PID/maps
Initial commit of aufs for v3.9.
By taking another approach to support mmap(2),
aufs adds a new member called vm_prfile in struct
vm_area_struct (and struct vm_region). The original vm_file points to
the file on the branch fs in order to handle everything correctly as
usual. The new vm_prfile points to a virtual file in aufs, and the
show-functions in procfs refers to vm_prfile if it is set.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>