aufs: for linux-3.12, add non-FMODE_WRITE-opened file to the list 2/2
It is important for aufs to add non-FMODE_WRITE-opened files to
sb->s_files, particularly the branch management.
In aufs branch management, all opened files are tested whether it is
mmap()ed or not. This is done regardless the is opened with FMODE_WRITE
or not.
So add it with aufs.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>