aufs: for linux-3.15, support for FS_DELETE in NFS silly-rename
By the commit
f7be728 2014-03-17 nfs: emit a fsnotify_nameremove call in sillyrename codepath
NFS fires the event by itself (while the event is fired by VFS usually).
Additionally NFS lets kworker to fire. This is an unexpected situation
for aufs since aufs handles the event asynchronously which means calling
kworker. But this will not cause a deadlock since aufs lets kworker
asynchronously (if aufs made it synchronously, then a deadlock would
happen).

Anyway a notify from kworker is unexpected for aufs and it produces a
warning when CONFIG_AUFS_DEBUG is enabled. This commit allows the nested
kworker call when the target is on NFS and the event is FS_DELETE, but a
minor warning is left.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>