aufs: support for weird dir nlink
This is mainly for btrfs branch.
Dan Kegel and Michael Johnson - MJ reported an ESTALE happens
unexpectedly when the branch is btrfs. MJ also posted a simple
reproducible script.
I could not reproduce the problem on my test environment, but I could
find the dir link count is 1 on btrfs. As far as I know, every
filesystem sets 2 to the newly created dir's link count. If I remember
correctly, it is due to a historical reason. I never say it is wrong,
since it is legal that the link count of a newly created dir is 1.

I developed a patch as a workaround, and fortunately MJ repeated the
test with the patch. Finally he could not reproduce the problem.

I believe the problem surely happens when the writable branch is btrfs,
although I could not reproduce the exact same problem. This commit fixes
a similar problem on btrfs, but is not confirmed that Dan and MJ's
problem is surely fixed.

See-also: http://www.mail-archive.com/aufs-users@lists.sourceforge.net/msg04979.html and its thread
Reported-by: Dan Kegel <dank@kegel.com>
Reported-by: Michael Johnson - MJ <mj@revmj.com>
Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>