aufs: copyup the timestamp from the lower branch
Basically the timestamps of aufs directory simply follows the topmost
entry, but it makes some user-space applications not to work well who
monitors the directory timestamps. When there are multiple writable
branches, aufs may choose the lower branch following the specified
policy to create or to copy-up. It means that the dir timestamp on the
topmost writable branch will not be updated, and aufs dir timestamp
either.
In order to work the monitoring applications correctly, this commit
registers an asynchronous work to copy-up the timestamp from lower to
topmost branch and aufs directory.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>