aufs: bugfix, check XINO before unlinking
Theoretically there is a very small chance for users to unlink the XINO
file. It is between its internal creation and internal deletion. Of
course such possibility is very small, but if it happens, the internal
deletion will trigger a problem. It is a simple "unlinking a file with
nlink == 0" problem. Let's check i_nlink before unlinking XINO file.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>