aufs: minor, convert radix_tree for si_pid to bitmap array
Hoping a little better performance when pid_max is customized as much
greater than default.
But I don't like the kcalloc loop approach.

If we could get a notify when pid_max is changed, then we might be able
to utilize the memory region a little better. But it is not a big deal.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>