aufs: bugfix, half-bad-inode
There was a regression in the commit
	caeaab8 2016-05-19 aufs: minor, simplify the support for the bad inodes

When  a race happens in VFS iget_locked(), it calls
destroy_inode(). Which means aufs_destroy_inode() is called for an inode
which is not fully set its info (aufs inode info), eg. half bad inode.
Here this commit handles such special case.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>