Revert "aufs: for linux-4.0, tiny, generic_update_time()"
This reverts commit e0d3c30e44f7f195329ea386607f30a1a1bfde94.

By the commit in linux-v4.0
	0ae45f6 2015-02-05 vfs: add support for a lazytime mount option
default_update_time() was introduced, and by the commit in aufs4.0
	e0d3c30 2015-03-03 aufs: for linux-4.0, tiny, generic_update_time()
aufs replaced a function (VFS:update_time()) which aufs used to call by
a new function generic_update_time().

But it was wrong. Aufs should keep calling VFS:update_time() in order to
support FS's ->update_time().

In the real world, ->update_time() is less popular and only btrfs and
xfs are implementing.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>